feat(orderbook): don't match same orders twice
This temporarily removes peer orders from the order book when they fail
swaps during the matching routine for a locally placed order. Placed
orders that fail any swaps will retry matching for the quantity that
failed, and previously this meant they may match again with the same
order resulting in a swap that will likely fail yet again.

This ensures that the two orders won't be matched again and that any
other orders in the order book with a matching price would be tried
instead. Once matching is complete, the order is added back to the
order book so it may be swapped again in the future by which time the
 reason for the swap failure may be resolved.

This replaces the `debug.testing` behavior for order matching.

Closes #1595.