fix(connext): async getbalance & error handling
This improves the parallelization and error handling of the `GetBalance`
call particularly as it pertains to Connext. Behind the scenes, Connext
has a single call that returns both on chain and off chain balance,
unlike lnd where the two calls are separate. xud was calling this twice,
which puts unnecessary load on Connext.

Instead, Connext now tracks when there is already an ongoing balance
request that it is waiting for a response for. Any `getBalance` calls
that happen while there is this wait will listen for the same response
rather than making separate calls.

Also, the error handling is changed for the `GetBalance` call such that
if we get errors for only some of the currencies for which we are
checking balances, we still return the balances for currencies which
did not give us any errors. This means that a connext error will not
cause the entire `GetBalance` call to fail when there are active lnd
currencies.

Fixes #1572.