feat(cli): print "no more matches" for place order
This prints a message when placing orders via the cli when there are
no more matches in the order book and a remainder of our order will be
discarded. For limit orders, we print a message for the remainder to
enter the order book in cases where we do not fully match right away.
However there was no such message for market and IOC orders.

A similar message is also printed to the xud logs.

Closes #1656.