fix(swaps): no remote fail when sending payment
This prevents the maker from failing a swap due to a swap failed packet
received from the taker if the maker has already started sending its
payment. This is because the taker may still claim the payment, and the
maker needs to continue monitoring its outgoing payment until it is
certain that the payment cannot be claimed.

Fixes #1749.