feat(lnd): handle delayed macaroon creation
This improves the initialization flow for LndClient when creating a new
wallet in lnd. Sometimes `admin.macaroon` is not created immediately
after calling `InitWallet` on lnd. Previously, this could lead to us
being stuck in a loop on `awaitWalletInit` immediately after creating
or restoring a wallet in lnd. If there are no macaroons by the first
time we attempt to verify the connection, we treated it as if lnd had
never received an `InitWallet` command and wait for that command to
happen, however that call never comes since it had already happened.

Instead, this adds new logic to detect the above scenario and skip
ahead to simply wait for `admin.macaroon` to be written to disk and
made available for lnd calls by xud.