fix(p2p): handle multiple socket errors
This changes the peer socket error handling to use `on` instead of `once`
when listening for `error` events to ensure that an unexpected second
`error` event doesn't crash xud after the `once` handler stops listening.
Instead we use `on` to listen to all error events and then manually
remove all listeners from the socket after it's destroyed.

Fixes #1773