feat(orderbook): prevent sub-satoshi order & match
This prevents xud from accepting and matching orders that would result
in less than a full satoshi from being exchanged for both currencies
involved in the swap. It does this three ways:

1. Preventing limit orders from being placed locally that do not meet
the satoshi requirement.

2. Preventing peer orders from being added to the order book if they do
not meet the satoshi requirement.

3. Preventing matches that do not meet the satoshi requirement, this is
needed for handling locally placed market orders since they do not have
a price assigned to them until they match with an order.

It is still possible for "dust" orders that cannot be traded to get
stuck in the orderbook if they are partially removed or matched leaving
a tiny fraction behind. A follow-up PR can automatically remove these.

Closes #1594.