feat(cli): print no more matches on partial mkt
This ensures that a "no more matches" message is printed by the cli
when a market order is only partially matched due to no more orders in
the order book. It does this by tracking the remaining quantity.
Previously, a message would only get printed if the matching ended on a
failed swap, but didn't cover the case where there were simply no more
orders available in the order book.

Closes #1596.