fix: restrict reputation events in non-strict mode
This ensures we only add severe and manual reputation events to peers
when not running in strict mode by filtering both events that arise from
the peer itself as well as those originating from other sources, such
as failed swaps. Previously we only checked on events arising from the
peer, which, for example, would include packet delays but not a failure
to find routes.

Fixes #1802