test(sim): handle no force close for taker
This modifies the simulation tests to account for the maker properly
canceling the invoice for its incoming payment once its outgoing payment
htlc expires. Rather than expecting the channel with the taker to maker
payment to expire and force close, it expects the invoice to be canceled
and for the channel to remain open. It ensures that the balances on the
channel do not change.