refactor(swaps): relax faildeal complete check
This relaxes the assertion in `failDeal` to ensure we don't "fail" a
deal that has already been completed. Previously such a case would crash
xud, now it only prints an error message and does nothing. Although this
shouldn't happen and indicates a flaw with xud code should it arise, it
is not a critical error and does not risk funds or indicate that xud is
inoperable.