fix(connext): prevent duplicate htlcAccepted evts (#1854)
* fix(connext): prevent duplicate htlcAccepted evts

This prevents trying to settle an invoice more than once in case Connext
sends us duplicate transfer received events. Previously, ConnextClient
could emit the `htlcAccepted` event multiple times, resulting in one
successful settleInvoice call and at least one failed call, which would
then attempt to fail a completed deal and crash xud.

Closes #1851.

* refactor(swaps): relax faildeal complete check

This relaxes the assertion in `failDeal` to ensure we don't "fail" a
deal that has already been completed. Previously such a case would crash
xud, now it only prints an error message and does nothing. Although this
shouldn't happen and indicates a flaw with xud code should it arise, it
is not a critical error and does not risk funds or indicate that xud is
inoperable.