feat(p2p): don't log empty order packets
This skips the log statement that logs when a peer sends us zero orders
for a given trading pair, as this is more or less a non-event and is
currently logged fairly frequently. It also lowers the level of logging
when we do get orders from `verbose` to `debug` as it is a common and
not particularly important event.