fix: don't activate unsupported pairs with peers
This ensures that, even when not running in strict mode, we only
activate trading pairs with peers when we support both currencies of the
pair and the trading pair itself locally. Previously we would activate
every trading pair, like BTC/LTC when we don't have a swap client for
LTC, resulting in us requesting and receiving order updates for trading
pairs that are irrelevent to us.