feat(connext): request collateral for order amount
This adds a check when a new order is receiving tokens via connext that
ensures we have sufficient collateral from the connext node to fulfill
the order. If not, we fail the order request and request the additional
collateral necessary. This involves the following changes:

1. The Connext client now tracks the inbound node collateral for each
currency and refreshes this value every time it queries for channel
balances and every time a collateral request completes.

2. Before placing a new order that is receiving Connext tokens we check
the inbound amount against the available collateral for that currency.
In the case of market orders, we use the best available matching price
in the order book to calculate inbound & outbound amounts.

3. If the collateral is insufficient, the order will be rejected and a
collateral request will be performed for the necessary capacity plus a
5% buffer.

4. Any other collateral checks while we're awaiting a response for a
collateral request will not perform an additional collateral request so
as to prevent duplicate calls.

5. Traders may repeat their order request and it will be accepted once
sufficient collateral to complete the trade is acquired.

6. Collateral is only requested upon depositing funds to a Connext
channel according to hardcoded collateral request minimums per currency.

There are also several hardcoded collateral request minimums. If the
capacity shortage is smaller than these minimums, the minimum amount
will be requested instead.

Closes #1845.