fix(orderbook): don't remove 0 quantity w/ hold
This fixes a bug whereby attempting to remove an order that is entirely
on hold would throw an error due to trying to remove zero quantity from
an order. Instead, if the entire order is on hold we skip removing any
portion that is not on hold (since there is none) and wait for the hold
to be lifted and remove any quantity.