feat: derive child seed for swap clients
This change uses separate mnemonics and seeds for each swap client that
are unique to that client, rather than using a single seed and mnemonic
across all clients. To accomplish this, a new method is added to the
seedutil tool to derive a child mnemonic from a provided mnemonic. This
is done by extracting the 16 bytes of entropy from the provided seed,
concatenating the entropy with the name of the client (e.g. "LND-BTC")
and then hashing the concatenation and taking 16 bytes from the
resulting hash.

Closes #1576.