feat(rpc): reserved order amount for TradingLimits
This exposes the amounts reserved for open orders on the `TradingLimits`
call and subtracts reserved amounts from the `maxSell` and `maxBuy`
values returned by the call.

It also removes reserved order amounts from GetBalance - reversing #1584.

The logic to calculate max sell and buys is as follows:

1. Sum up all outbound and inbound balances across all channels, as well
as the maximum balances for a single channel, as we still aren't
expecting to use multi-path payments. For Connext there is a single
channel so we only have a single inbound & outbound value per currency.

2. Subtract reserved order amounts from the total inbound and outbound
balances calculated above. E.g. if we have outbound capacity of 1.2 ETH,
but we have an order to sell 0.7 ETH, our max sell would be 0.5 ETH.

3. Return the smaller of the amount calculated in step 2 or the maximum
single channel balance calculated in step 1.

Resolves #1678.