feat(rpc): runtime addcurrency for lnd & connext
This PR allows for adding Connext and Lnd currencies via the
`AddCurrency` rpc call and have them available for trading immediately
without requiring a restart of xud.

In the case of Connext it merely registers the tokenaddress and currency
symbol for the new currency with the existing ConnextClient, as we had
previously done with Raiden.

In the case of Lnd, we read from the configuration and use that to
instantiate and initialize a new `LndClient` for the specified currency.

Closes #1111.

Co-authored-by: Le Premier Homme <interjoint1@gmail.com>