feat(rpc): InitSwapClient
This allows for xud to initialize a swap client after xud itself has
already been initialized using a new rpc command `InitSwapClient`. The
swap client must be configured properly and running before it can be
initialized. The procedure for initializing is as follows.

1. A swap client type and (optionally) a currency is specified by the
`InitSwapClient` request.
2. Xud extracts the deciphered seed from its decrypted node key, this
contains the entropy used to generate the master seed/mnemonic.
3. Xud uses the seedutil to convert the deciphered seed back to a
mnemonic. For Lnd, it uses this mnemonic to initialize the wallet.

Closes #1929.