fix(p2p): fix stall timer checks
This fixes a regression introduced in #1994 that prevents enabling the
stall timer for peers that checks whether peers are not responding to
our packets in a reasonable timeframe and disconnecting those that are
not. Without these checks, peers that have gone offline may remain in
the list of connected peers, resulting in unexpected behavior and also
preventing those peers from establishing new connections with us once
they come back online.

Closes #2010.