refactor!: migrate to @grpc/grpc-js
This migrates the `grpc` package to the pure javascript `grpc-js`
implementation. `grpc-js` is the more actively maintained package and
the preferred implementation by Google projects, and it reduces the
dependency on and overhead from compilers like `node-pre-gyp`. `grpc`
is now deprecated in npm.

This also updates `grpc-tools` and tweaks the way grpc code is generated
from `.proto` files to be compatible with `grpc-js`.

Lastly this removes the grpc web proxy from xud. It is not compatible
with `grpc-js` and it is currently not used for any purpose including
testing or development. As the old approach relied on a custom fork of
an old, unmaintained package to create the proxy, we plan to reimplement
a web proxy with a new approach using an add-on or module that runs the
proxy in a separate process.

BREAKING CHANGE: Removes grpc web proxy from xud.

Co-authored-by: LePremierHomme <interjoint3@gmail.com>