feat(rpc): OrderBook rpc call
This creates a new `OrderBook` rpc call to return a depth chart for the
xud order book. This repurposes the logic that previously existed only
in the `xucli orderbook` call which groups orders into buckets according
to rounded price and moves that to the rpc layer, so that consumers of
the API can get an xud depth chart without needing to reimplement the
depth chart/bucket logic.

Closes #2048.